Start Refactoring Swagger Test
