+++ composer.json
@@ -13,7 +13,8 @@
   "require": {
     "guzzlehttp/guzzle": "6.*",
     "byjg/uri": "1.0.*",
-    "php": ">=5.6"
+    "php": ">=5.6",
+    "ext-json": "*"
   },
   "require-dev": {
     "phpunit/phpunit": "5.7.*|7.4.*"
