+++ src/SwaggerSchema.php
@@ -2,6 +2,7 @@
 
 namespace ByJG\Swagger;
 
+use ByJG\Swagger\Base\Schema;
 use ByJG\Swagger\Exception\DefinitionNotFoundException;
 use ByJG\Swagger\Exception\HttpMethodNotFoundException;
 use ByJG\Swagger\Exception\InvalidDefinitionException;
@@ -9,54 +10,15 @@ use ByJG\Swagger\Exception\NotMatchedException;
 use ByJG\Swagger\Exception\PathNotFoundException;
 use ByJG\Util\Uri;
 
-class SwaggerSchema
+class SwaggerSchema extends Schema
 {
-    protected $jsonFile;
-    protected $allowNullValues;
-    protected $specificationVersion;
-
-    protected $serverVariables = [];
-
-    const SWAGGER_PATHS = "paths";
-    const SWAGGER_PARAMETERS = "parameters";
-    const SWAGGER_COMPONENTS = "components";
-
     public function __construct($jsonFile, $allowNullValues = false)
     {
-        $this->jsonFile = json_decode($jsonFile, true);
+        if (!is_array($jsonFile)) {
+            $jsonFile = json_decode($jsonFile, true);
+        }
+        $this->jsonFile = $jsonFile;
         $this->allowNullValues = (bool) $allowNullValues;
-        $this->specificationVersion = isset($this->jsonFile['swagger']) ? '2' : '3';
-    }
-
-    /**
-     * Returns the major specification version
-     * @return string
-     */
-    public function getSpecificationVersion()
-    {
-        return $this->specificationVersion;
-    }
-
-    public function getServerUrl()
-    {
-        if (!isset($this->jsonFile['servers'])) {
-            return '';
-        }
-        $serverUrl = $this->jsonFile['servers'][0]['url'];
-
-        if (isset($this->jsonFile['servers'][0]['variables'])) {
-            foreach ($this->jsonFile['servers'][0]['variables'] as $var => $value) {
-                if (!isset($this->serverVariables[$var])) {
-                    $this->serverVariables[$var] = $value['default'];
-                }
-            }
-        }
-
-        foreach ($this->serverVariables as $var => $value) {
-            $serverUrl = preg_replace("/\{$var\}/", $value, $serverUrl);
-        }
-
-        return $serverUrl;
     }
 
     public function getHttpSchema()
@@ -71,74 +33,9 @@ class SwaggerSchema
 
     public function getBasePath()
     {
-        if ($this->getSpecificationVersion() === '3') {
-            $uriServer = new Uri($this->getServerUrl());
-            return $uriServer->getPath();
-        }
-
         return isset($this->jsonFile['basePath']) ? $this->jsonFile['basePath'] : '';
     }
 
-    /**
-     * @param $path
-     * @param $method
-     * @return mixed
-     * @throws DefinitionNotFoundException
-     * @throws HttpMethodNotFoundException
-     * @throws InvalidDefinitionException
-     * @throws NotMatchedException
-     * @throws PathNotFoundException
-     */
-    public function getPathDefinition($path, $method)
-    {
-        $method = strtolower($method);
-
-        $path = preg_replace('~^' . $this->getBasePath() . '~', '', $path);
-
-        $uri = new Uri($path);
-
-        // Try direct match
-        if (isset($this->jsonFile[self::SWAGGER_PATHS][$uri->getPath()])) {
-            if (isset($this->jsonFile[self::SWAGGER_PATHS][$uri->getPath()][$method])) {
-                return $this->jsonFile[self::SWAGGER_PATHS][$uri->getPath()][$method];
-            }
-            throw new HttpMethodNotFoundException("The http method '$method' not found in '$path'");
-        }
-
-        // Try inline parameter
-        foreach (array_keys($this->jsonFile[self::SWAGGER_PATHS]) as $pathItem) {
-            if (strpos($pathItem, '{') === false) {
-                continue;
-            }
-
-            $pathItemPattern = '~^' . preg_replace('~{(.*?)}~', '(?<\1>[^/]+)', $pathItem) . '$~';
-
-            $matches = [];
-            if (preg_match($pathItemPattern, $uri->getPath(), $matches)) {
-                $pathDef = $this->jsonFile[self::SWAGGER_PATHS][$pathItem];
-                if (!isset($pathDef[$method])) {
-                    throw new HttpMethodNotFoundException("The http method '$method' not found in '$path'");
-                }
-
-                $parametersPathMethod = [];
-                $parametersPath = [];
-
-                if (isset($pathDef[$method][self::SWAGGER_PARAMETERS])) {
-                    $parametersPathMethod = $pathDef[$method][self::SWAGGER_PARAMETERS];
-                }
-
-                if (isset($pathDef[self::SWAGGER_PARAMETERS])) {
-                    $parametersPath = $pathDef[self::SWAGGER_PARAMETERS];
-                }
-
-                $this->validateArguments('path', array_merge($parametersPathMethod, $parametersPath), $matches);
-
-                return $pathDef[$method];
-            }
-        }
-
-        throw new PathNotFoundException('Path "' . $path . '" not found');
-    }
 
     /**
      * @param $parameterIn
@@ -148,33 +45,8 @@ class SwaggerSchema
      * @throws InvalidDefinitionException
      * @throws NotMatchedException
      */
-    private function validateArguments($parameterIn, $parameters, $arguments)
+    protected function validateArguments($parameterIn, $parameters, $arguments)
     {
-        if ($this->getSpecificationVersion() === '3') {
-            foreach ($parameters as $parameter) {
-                if (isset($parameter['$ref'])) {
-                    $paramParts = explode("/", $parameter['$ref']);
-                    if (count($paramParts) != 4 || $paramParts[0] != "#" || $paramParts[1] != self::SWAGGER_COMPONENTS || $paramParts[2] != self::SWAGGER_PARAMETERS) {
-                        throw new InvalidDefinitionException(
-                            "Not get the reference in the expected format #/components/parameters/<NAME>"
-                        );
-                    }
-                    if (!isset($this->jsonFile[self::SWAGGER_COMPONENTS][self::SWAGGER_PARAMETERS][$paramParts[3]])) {
-                        throw new DefinitionNotFoundException(
-                            "Not find reference #/components/parameters/${paramParts[3]}"
-                        );
-                    }
-                    $parameter = $this->jsonFile[self::SWAGGER_COMPONENTS][self::SWAGGER_PARAMETERS][$paramParts[3]];
-                }
-                if ($parameter['in'] === $parameterIn &&
-                    $parameter['schema']['type'] === "integer"
-                    && filter_var($arguments[$parameter['name']], FILTER_VALIDATE_INT) === false) {
-                    throw new NotMatchedException('Path expected an integer value');
-                }
-            }
-            return;
-        }
-
         foreach ($parameters as $parameter) {
             if ($parameter['in'] === $parameterIn
                 && $parameter['type'] === "integer"
@@ -193,18 +65,6 @@ class SwaggerSchema
     public function getDefintion($name)
     {
         $nameParts = explode('/', $name);
-
-        if ($this->getSpecificationVersion() === '3') {
-            if (count($nameParts) < 4 || $nameParts[0] !== '#') {
-                throw new InvalidDefinitionException('Invalid Component');
-            }
-
-            if (!isset($this->jsonFile[$nameParts[1]][$nameParts[2]][$nameParts[3]])) {
-                throw new DefinitionNotFoundException("Component'$name' not found");
-            }
-
-            return $this->jsonFile[$nameParts[1]][$nameParts[2]][$nameParts[3]];
-        }
 
         if (count($nameParts) < 3 || $nameParts[0] !== '#') {
             throw new InvalidDefinitionException('Invalid Definition');
@@ -231,50 +91,10 @@ class SwaggerSchema
     {
         $structure = $this->getPathDefinition($path, $method);
 
-        if($this->getSpecificationVersion() === '3') {
-            if (!isset($structure['requestBody'])) {
-                return new SwaggerRequestBody($this, "$method $path", []);
-            }
-            return new SwaggerRequestBody($this, "$method $path", $structure['requestBody']);
-        }
-
         if (!isset($structure[self::SWAGGER_PARAMETERS])) {
             return new SwaggerRequestBody($this, "$method $path", []);
         }
         return new SwaggerRequestBody($this, "$method $path", $structure[self::SWAGGER_PARAMETERS]);
-    }
-
-    /**
-     * @param $path
-     * @param $method
-     * @param $status
-     * @return SwaggerResponseBody
-     * @throws HttpMethodNotFoundException
-     * @throws InvalidDefinitionException
-     * @throws NotMatchedException
-     * @throws PathNotFoundException
-     * @throws DefinitionNotFoundException
-     */
-    public function getResponseParameters($path, $method, $status)
-    {
-        $structure = $this->getPathDefinition($path, $method);
-
-        if (!isset($structure['responses'][$status])) {
-            throw new InvalidDefinitionException("Could not found status code '$status' in '$path' and '$method'");
-        }
-
-        return new SwaggerResponseBody($this, "$method $status $path", $structure['responses'][$status]);
-    }
-
-    /**
-     * OpenApi 2.0 doesn't describe null values, so this flag defines,
-     * if match is ok when one of property
-     *
-     * @return bool
-     */
-    public function isAllowNullValues()
-    {
-        return $this->allowNullValues;
     }
 
     /**
